daemon: add host IPs to IPCache
Egress to host did not work with policy containing toEndpoints "reserved:host"
label because previously, only endpoint IPs were added to the local IPCache in
each cilium-agent. To allow for label-based selecting of the host, we need to
add the host IPs to the IPCache so lookup in the datapath succeeds when mapping
the IP of the host to its identity.

Signed-off by: Ian Vernon <ian@cilium.io>